Incorporate my own suggestions.
