+++ sparse/coo/common.py
@@ -8,7 +8,7 @@ import scipy.sparse
 from ..sparse_array import SparseArray
 from ..compatibility import range
 from ..utils import isscalar
-from .indexing import _normalize_axis
+from sparse.utils import _normalize_axis
 
 
 def asCOO(x, name='asCOO', check=True):
